chore: replace existing fisk and poisson dist and add authors
